+++ app/error/AdapterError.go
