+++ app/http/response/transformers/SampleTransformer.go
@@ -1,7 +1,7 @@
 package transformers
 
 import (
-	"github.com/kosatnkn/catalyst/app/http/response"
+	"github.com/kosatnkn/catalyst/app/http/response/transformers"
 	"github.com/kosatnkn/catalyst/domain/entities"
 )
 
@@ -12,7 +12,7 @@ type SampleTransformer struct {
 }
 
 // NewSampleTransformer creates a new instance of the transformer.
-func NewSampleTransformer() response.TransformerInterface {
+func NewSampleTransformer() transformers.TransformerInterface {
 	return &SampleTransformer{}
 }
