+++ externals/adapters/MySQLAdapter.go
@@ -11,7 +11,7 @@ import (
 	_ "github.com/go-sql-driver/mysql"
 
 	"github.com/kosatnkn/catalyst/app/config"
-	errTypes "github.com/kosatnkn/catalyst/app/error/types"
+	externalErrs "github.com/kosatnkn/catalyst/app/error"
 	"github.com/kosatnkn/catalyst/domain/boundary/adapters"
 	"github.com/kosatnkn/catalyst/domain/globals"
 )
@@ -125,7 +125,7 @@ func (a *MySQLAdapter) reorderParameters(params map[string]interface{}, namedPar
 		paramValue, isParamExist := params[param]
 
 		if !isParamExist {
-			return nil, errTypes.NewAdapterError(fmt.Sprintf("parameter '%s' is missing", param), 100, "")
+			return nil, externalErrs.NewAdapterError(fmt.Sprintf("parameter '%s' is missing", param), 100, "")
 		}
 
 		reorderedParams = append(reorderedParams, paramValue)
