+++ externals/adapters/PostgresAdapter.go
@@ -11,7 +11,7 @@ import (
 	_ "github.com/lib/pq"
 
 	"github.com/kosatnkn/catalyst/app/config"
-	errTypes "github.com/kosatnkn/catalyst/app/error/types"
+	externalErrs "github.com/kosatnkn/catalyst/app/error"
 	"github.com/kosatnkn/catalyst/domain/boundary/adapters"
 	"github.com/kosatnkn/catalyst/domain/globals"
 )
@@ -132,7 +132,7 @@ func (a *PostgresAdapter) reorderParameters(params map[string]interface{}, named
 		paramValue, isParamExist := params[param]
 
 		if !isParamExist {
-			return nil, errTypes.NewAdapterError(fmt.Sprintf("parameter '%s' is missing", param), 100, "")
+			return nil, externalErrs.NewAdapterError(fmt.Sprintf("parameter '%s' is missing", param), 100, "")
 		}
 
 		reorderedParams = append(reorderedParams, paramValue)
